+++ data/taskList.txt
@@ -1,5 +1,7 @@
-T | 0 | buy things
-T | 0 | buy something
-D | 0 | return book | 17/11/2019 1230
-E | 1 | attend wedding | 22/11/2019 1830
-T | 0 | buy computer
+T | 0 | buy food
+T | 1 | buy computer
+T | 0 | do housework
+D | 1 | excercise | 12/12/2019 1100
+T | 0 | borrow book
+T | 0 | buy book
+D | 0 | return book | 19/11/2019 1230
