+++ src/main/java/command/ByeCommand.java
@@ -1,3 +1,10 @@
+package command;
+
+import exceptions.DukeException;
+import task.Storage;
+import task.TaskList;
+import ui.UI;
+
 public class ByeCommand extends Command {
     public static final String INPUT_WORD = "bye";
     public static final String MESSAGE_BYE_ACKNOWLEDGEMENT = "Bye. Hope to see you again soon!";
