+++ src/main/java/task/Event.java
@@ -1,3 +1,5 @@
+package task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 //import java.util.Date;
