+++ src/main/java/task/Storage.java
@@ -1,3 +1,7 @@
+package task;
+
+import exceptions.DukeException;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
