+++ src/main/java/ui/UI.java
@@ -1,3 +1,5 @@
+package ui;
+
 import java.util.Scanner;
         import java.io.InputStream;
